requirements.txt
matplotlib==3.3.2
keras_nightly==2.5.0.dev2021032900
Flask==2.0.1
numpy==1.16.2
h5py==3.1.0
pymongo==3.11.4
Flask_PyMongo==2.3.0
opencv_python==4.5.2.54
tensorflow_gpu==2.5.0
scipy==1.2.1
requests==2.25.1
pandas==1.1.5
keras==2.4.3
Pillow==8.2.0
